deprecate convert_id_name_community function
